Yices: Upgrade to gmp-6.3.0
Previously, we were using `gmp-6.2.1`, which was susceptible to a serious bug
on ARM64 that manifested as Yices (a tool which statically links against `gmp`)
to segfault on ARM64. See https://gitlab.haskell.org/ghc/ghc/-/issues/22497 for
a very similar bug that arose in GHC.

This fixes the issue by upgrading to `gmp-6.3.0`, which includes a fix for this
bug. As a result, this also fixes #58.