Fix the intel MacOS 15.x yices build.
Problem was:
   - gmp for some reason built itself incorrectly and produced a
     library full of illegal text relocations;
   - libpoly's configure script detected that the gmp library didn't
     work, but instead of failing decided to just blithely continue
     and not link with the broken library;
   - then the yices build fails on the libpoly library.

Fix it by passing --with-pic when on intel MacOS; this apparently
causes gmp to build only PIC libraries, which don't have text
relocations.

It's possible that gmp was always building a library full of text
relocations and it only started failing on the newer MacOS; it's also
possible that something in its configury is broken; it's not obvious
from casual inspection and there's a lot of goo there.