Migrate from cvc5-1.1.1 to 1.3.1
This is primarily motivated by bringing in the changes from
https://github.com/cvc5/cvc5/pull/11861, which brings in a newer version of
`libpoly` that is compatible with recent versions of Apple `clang++`. (See
https://github.com/SRI-CSL/libpoly/pull/85 for the relevant `libpoly` patch.)
This, in turn, allows `libpoly` to build with `macos-15` CI runners.

Note that recent versions of `cvc5` no longer require patching out the use of
the `ld.gold` linker on Windows, so we can remove the corresponding
`what4-solvers` patch that does this.